refactor: change amount type

+++ doners-backend/src/main/java/com/doners/donersbackend/domain/dao/donation/Donation.java
@@ -39,7 +39,7 @@ public class Donation extends BaseEntity {
     private String description;
 
     @Column(name = "donation_amount")
-    private long amount;
+    private double amount;
 
     @Column(name = "donation_start_date")
     private LocalDate startDate;
